Reset dnf module virt before switching it to virt:av
To prevent this error:
The operation would result in switching of module 'virt' stream 'rhel'
to stream 'av'
Error: It is not possible to switch enabled streams of a module.
It is recommended to remove all installed content from the module,
and reset the module using 'dnf module reset <module_name>' command.
After you reset the module, you can install the other stream.