Mark unbound PersistentVolumeClaim as healthy
The built-in Argo CD health check for PersistentVolumeClaim will not
mark a PVC as healthy until it is bound. PVCs that are WaitForFirstConsumer
will not be marked as healthy until the PV is created and bound to them.

This commit changes Argo CD health check so that an unbound PVC is
deemed healthy.